refcator : seperated components and style issues
